Use resource id from build.gradle instead of packageName
getPackageName will return the identifier string that contains the
applicationIdSuffix. Since the suffix changes with the build env, the
MainActivity can only be launched using the root applicationId.

For this to work, the app/build.gradle file will have to expose a
resValue for the default applicationId:

`resValue "string", "build_config_package", "com.example.foo"`